feat: Add homepage
A HomeController with an index action has been created to define the application's root route. The HomeController is associated with a new home helper module and a corresponding index view. The changes are made to provide a home page for our web application. They establish the basic structure for additional features and routing. This is the first part of our application layout work.